Graphs as nodes:
Graphs can now be inserted as nodes in other graphs.
- Use the In-/Output nodes found in Flow/Graph/ to specify pins for the graph.
- Graphs are added to the node list through special install functions, see startup.py.

Executing graphs through the console has also been made a bit smoother, see flow.run():
flow.run() takes a graph or a filepath to a graph as the first argument, then it takes keyword-arguments specifying the input values. Returnes is a dictionary with all the outputs.